+++ RSNNS/src/cc_learn.cpp
@@ -110,7 +110,7 @@ krui_err SnnsCLib::cc_initVariables(float* ParameterInArray,
 	cc_Parameter[i]=ParameterInArray[22+i];
     } 
     cc_fastmode=(int)ParameterInArray[27];
-    cc_fse=((LEARNING_FUNCTION==BACKPROP)? PARAM3 : 0.1);
+    cc_fse=((LEARNING_FUNCTION==BACKPROP)? PARAM3 : 0.1f);
 
     KernelErrorCode=cc_InitModificationVariables();
     ERROR_CHECK;
@@ -153,9 +153,9 @@ float SnnsCLib::cc_calculateCorrelation(int StartPattern, int EndPattern, int co
 {
 
     int s,o,n;
-    double highScore, unchangedhighScore, scoreBuffer;
-    float bestSpecialUnitScore = -0.1 ,
-	bestSpecialUnitUnchangedScore=0.0;
+    float highScore, unchangedhighScore, scoreBuffer;
+    float bestSpecialUnitScore = -0.1f ,
+	bestSpecialUnitUnchangedScore=0.0f;
     struct Unit *SpecialUnitPtr = NULL,*OutputUnitPtr;
     int start, end;
 
@@ -165,7 +165,7 @@ float SnnsCLib::cc_calculateCorrelation(int StartPattern, int EndPattern, int co
     //if(cc_printOnOff)  printf("Cycle %d ",counter);
 
     FOR_ALL_SPECIAL_UNITS(SpecialUnitPtr,s) {
-	unchangedhighScore = 0.0;
+	unchangedhighScore = 0.0f;
 	FOR_ALL_OUTPUT_UNITS(OutputUnitPtr,o) {
 	    scoreBuffer = (CorBetweenSpecialActAndOutError[s][o] - 
 			   (MeanOutputUnitError[o] * SpecialUnitSumAct[s] )) / 
@@ -258,10 +258,10 @@ void SnnsCLib::cc_trainSpecialUnits(int maxNoOfCovarianceUpdateCycles,
 
     cc_printHeadline(const_cast<char*>("Training of the candidates"),LENGTH_HEADLINE);
     cc_calculateOutputUnitError(StartPattern,EndPattern);
-    if (SumSqError==0.0){
+    if (SumSqError==0.0f){
 	//Error = 
         cc_getErr(StartPattern,EndPattern); /* calc SumSqError */
-	if (SumSqError==0.0)
+	if (SumSqError==0.0f)
 	    return;
     }
     for (counter=0;counter<maxNoOfCovarianceUpdateCycles;counter++){
@@ -400,7 +400,7 @@ void SnnsCLib::cc_calculateOutputUnitError(int StartPattern,int EndPattern)
     int start, end,pat,sub,n;
   
     FOR_ALL_OUTPUT_UNITS(UnitPtr,o) 
-       OUTPUT_UNIT_SUM_ERROR[o] = 0.0;
+       OUTPUT_UNIT_SUM_ERROR[o] = 0.0f;
 
     cc_getPatternParameter(StartPattern,EndPattern,&start,&end,&n);
     ERROR_CHECK_WRC;
@@ -530,7 +530,7 @@ float SnnsCLib::cc_propagateOutputOnlineCase(int PatternNo, int sub_pat_no,
     int           dummy;
     float         lastChange;
 
-    sum_error = 0.0; 
+    sum_error = 0.0f; 
     out_pat = kr_getSubPatData(PatternNo,sub_pat_no,OUTPUT,NULL);
     ERROR_CHECK;
     FOR_ALL_OUTPUT_UNITS(OutputUnitPtr,dummy){
@@ -585,7 +585,7 @@ krui_err SnnsCLib::cc_propagateSpecial(int start,int end,int n,int counter,
         cc_getActivationsForActualPattern(p,start,&pat,&sub);
 
 	FOR_ALL_SPECIAL_UNITS(SpecialUnitPtr,s) {
-	    change = 0.0;
+	    change = 0.0f;
 	    SpecialUnitPtr->act = SpecialUnitAct[p][s];
 	    actPrime = (this->*SpecialUnitPtr->act_deriv_func)(SpecialUnitPtr);
 	    FOR_ALL_OUTPUT_UNITS(OutputUnitPtr,o) 
@@ -628,7 +628,7 @@ krui_err SnnsCLib::cc_propagateSpecialOnlineCase(int start,int end,int n,int cou
         cc_getActivationsForActualPattern(p,start,&pat,&sub);
 
 	FOR_ALL_SPECIAL_UNITS(SpecialUnitPtr,s) {
-	    change = 0.0;
+	    change = 0.0f;
 	    SpecialUnitPtr->act = SpecialUnitAct[p][s];
 	    actPrime = (this->*SpecialUnitPtr->act_deriv_func)(SpecialUnitPtr);
 	    FOR_ALL_OUTPUT_UNITS(OutputUnitPtr,o) 
@@ -925,7 +925,7 @@ krui_err SnnsCLib::cc_generateHiddenUnit(int GroupNo)
 	KernelErrorCode = krui_setCurrentUnit(CurrentUnit); 
 	ERROR_CHECK;
 
-	KernelErrorCode = krui_createLink(NewHiddenUnit,0.0);
+	KernelErrorCode = krui_createLink(NewHiddenUnit,0.0f);
 	ERROR_CHECK; 
     }       /* and change its Layer if necessary */
     KernelErrorCode = kr_topoSort(TOPOLOGICAL_CC);
